Moved MediaItemDoesNotExist to MediaMutations
