+++ layers/CMSSchema/packages/media-mutations/src/ObjectModels/MediaItemDoesNotExistErrorPayload.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace PoPCMSSchema\CustomPostMediaMutations\ObjectModels;
+namespace PoPCMSSchema\MediaMutations\ObjectModels;
 
 use PoPSchema\SchemaCommons\ObjectModels\AbstractErrorPayload;
