+++ layers/CMSSchema/packages/media-mutations/src/RelationalTypeDataLoaders/ObjectType/MediaItemDoesNotExistErrorPayloadObjectTypeDataLoader.php
@@ -2,9 +2,9 @@
 
 declare(strict_types=1);
 
-namespace PoPCMSSchema\CustomPostMediaMutations\RelationalTypeDataLoaders\ObjectType;
+namespace PoPCMSSchema\MediaMutations\RelationalTypeDataLoaders\ObjectType;
 
-use PoPCMSSchema\CustomPostMediaMutations\ObjectModels\MediaItemDoesNotExistErrorPayload;
+use PoPCMSSchema\MediaMutations\ObjectModels\MediaItemDoesNotExistErrorPayload;
 use PoP\ComponentModel\RelationalTypeDataLoaders\ObjectType\AbstractDictionaryObjectTypeDataLoader;
 
 class MediaItemDoesNotExistErrorPayloadObjectTypeDataLoader extends AbstractDictionaryObjectTypeDataLoader
